Se completa controlador de empresa

+++ src/main/java/com/jaded/sistemagestionproyectos/model/rol.java
@@ -8,7 +8,7 @@ public class rol {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long IdRole;
+    private Integer IdRole;
 
     @Column(length = 80)
     private String NombreRol;
